+++ include/config/CallbackEvents.h
@@ -1,0 +1,11 @@
+#ifndef CALLBACK_EVENTS
+#define CALLBACK_EVENTS
+
+#include <cstdint>
+
+enum CallbackEvents: int
+{
+    DRAW = 0
+};
+
+#endif